SparseConnectivityTracer

Documentation for SparseConnectivityTracer.

SparseConnectivityTracer.Tracer

SparseConnectivityTracer.connectivity

SparseConnectivityTracer.inputs

SparseConnectivityTracer.sortedinputs

SparseConnectivityTracer.tracer

API reference

SparseConnectivityTracer works by pushing a Number type called Tracer through generic functions:

SparseConnectivityTracer.Tracer — Type

Tracer(indexset) <: Number

Number type keeping track of input indices of previous computations.

See also the convenience constructor tracer.

Examples

julia> x = tracer(1, 2, 3)
Tracer(1, 2, 3)

julia> sin(x)
Tracer(1, 2, 3)

julia> 2 * x^3
Tracer(1, 2, 3)

julia> 0 * x   # Note: Tracer is strictly operator overloading...
Tracer(1, 2, 3)

julia> zero(x) # ...this can be overloaded
Tracer()

julia> y = tracer(3, 5)
Tracer(3, 5)

julia> x + y
Tracer(1, 2, 3, 5)

julia> x ^ y
Tracer(1, 2, 3, 5)

julia> M = rand(Tracer, 3, 2)
3×2 Matrix{Tracer}:
 Tracer()  Tracer()
 Tracer()  Tracer()
 Tracer()  Tracer()

julia> similar(M)
3×2 Matrix{Tracer}:
 Tracer()  Tracer()
 Tracer()  Tracer()
 Tracer()  Tracer()

julia> M * [x, y]
3-element Vector{Tracer}:
 Tracer(1, 2, 3, 5)
 Tracer(1, 2, 3, 5)
 Tracer(1, 2, 3, 5)

SparseConnectivityTracer.tracer — Function

tracer(index)
tracer(indices)

Convenience constructor for Tracer from input indices.

The resulting connectivity matrix can be extracted using connectivity:

SparseConnectivityTracer.connectivity — Function

connectivity(f, x)

Enumerates inputs x and primal outputs y=f(x) and returns sparse connectivity matrix C of size (m, n) where C[i, j] is true if the compute graph connects the i-th entry in y to the j-th entry in x.

or manually from individual Tracer outputs:

SparseConnectivityTracer.inputs — Function

inputs(tracer)

Return raw UInt64 input indices of a Tracer.

SparseConnectivityTracer.sortedinputs — Function

sortedinputs(tracer)
sortedinputs([T=Int], tracer)

Return sorted input indices of a Tracer.