API Reference

SparseConnectivityTracer.ConnectivityTracer

SparseConnectivityTracer.HessianTracer

SparseConnectivityTracer.JacobianTracer

SparseConnectivityTracer.TracerSparsityDetector

SparseConnectivityTracer.inputs

SparseConnectivityTracer.pattern

SparseConnectivityTracer.trace_input

SparseConnectivityTracer.tracer

Interface

SparseConnectivityTracer.pattern — Function

pattern(f, ConnectivityTracer{S}, x) where {S<:AbstractSet{<:Integer}}

Enumerates inputs x and primal outputs y = f(x) and returns sparse matrix C of size (m, n) where C[i, j] is true if the compute graph connects the i-th entry in y to the j-th entry in x.

pattern(f, JacobianTracer{S}, x) where {S<:AbstractSet{<:Integer}}

Computes the sparsity pattern of the Jacobian of y = f(x).

pattern(f, HessianTracer{S}, x) where {S<:AbstractSet{<:Integer}}

Computes the sparsity pattern of the Hessian of a scalar function y = f(x).

Examples

First order

julia> x = rand(3);

julia> f(x) = [x[1]^2, 2 * x[1] * x[2]^2, sin(x[3])];

julia> pattern(f, ConnectivityTracer{BitSet}, x)
3×3 SparseArrays.SparseMatrixCSC{Bool, UInt64} with 4 stored entries:
 1  ⋅  ⋅
 1  1  ⋅
 ⋅  ⋅  1

Second order

julia> x = rand(5);

julia> f(x) = x[1] + x[2]*x[3] + 1/x[4] + 1*x[5];

julia> pattern(f, HessianTracer{BitSet}, x)
5×5 SparseArrays.SparseMatrixCSC{Bool, UInt64} with 3 stored entries:
 ⋅  ⋅  ⋅  ⋅  ⋅
 ⋅  ⋅  1  ⋅  ⋅
 ⋅  1  ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅  1  ⋅
 ⋅  ⋅  ⋅  ⋅  ⋅

julia> g(x) = f(x) + x[2]^x[5];

julia> pattern(g, HessianTracer{BitSet}, x)
5×5 SparseArrays.SparseMatrixCSC{Bool, UInt64} with 7 stored entries:
 ⋅  ⋅  ⋅  ⋅  ⋅
 ⋅  1  1  ⋅  1
 ⋅  1  ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅  1  ⋅
 ⋅  1  ⋅  ⋅  1

pattern(f!, y, JacobianTracer{S}, x) where {S<:AbstractSet{<:Integer}}

Computes the sparsity pattern of the Jacobian of f!(y, x).

pattern(f!, y, ConnectivityTracer{S}, x) where {S<:AbstractSet{<:Integer}}

Enumerates inputs x and primal outputs y after f!(y, x) and returns sparse matrix C of size (m, n) where C[i, j] is true if the compute graph connects the i-th entry in y to the j-th entry in x.

SparseConnectivityTracer.TracerSparsityDetector — Type

TracerSparsityDetector <: ADTypes.AbstractSparsityDetector

Singleton struct for integration with the sparsity detection framework of ADTypes.jl.

Example

julia> using ADTypes, SparseConnectivityTracer

julia> ADTypes.jacobian_sparsity(diff, rand(4), TracerSparsityDetector())
3×4 SparseArrays.SparseMatrixCSC{Bool, UInt64} with 6 stored entries:
 1  1  ⋅  ⋅
 ⋅  1  1  ⋅
 ⋅  ⋅  1  1

julia> using ADTypes, SparseConnectivityTracer

julia> f(x) = x[1] + x[2]*x[3] + 1/x[4];

julia> ADTypes.hessian_sparsity(f, rand(4), TracerSparsityDetector())
4×4 SparseArrays.SparseMatrixCSC{Bool, UInt64} with 3 stored entries:
 ⋅  ⋅  ⋅  ⋅
 ⋅  ⋅  1  ⋅
 ⋅  1  ⋅  ⋅
 ⋅  ⋅  ⋅  1

Internals

SparseConnectivityTracer works by pushing Number types called tracers through generic functions. Currently, two tracer types are provided:

SparseConnectivityTracer.ConnectivityTracer — Type

ConnectivityTracer{S}(indexset) <: Number

Number type keeping track of input indices of previous computations. The provided set type S has to be an AbstractSet{<:Integer}, e.g. BitSet or Set{UInt64}.

See also the convenience constructor tracer. For a higher-level interface, refer to pattern.

SparseConnectivityTracer.JacobianTracer — Type

JacobianTracer{S}(indexset) <: Number

Number type keeping track of input indices of previous computations with non-zero derivatives. The provided set type S has to be an AbstractSet{<:Integer}, e.g. BitSet or Set{UInt64}.

See also the convenience constructor tracer. For a higher-level interface, refer to pattern.

SparseConnectivityTracer.HessianTracer — Type

HessianTracer{S}(indexset) <: Number

Number type keeping track of input indices of previous computations with non-zero first and second derivatives. The provided set type S has to be an AbstractSet{<:Integer}, e.g. BitSet or Set{UInt64}.

See also the convenience constructor tracer. For a higher-level interface, refer to pattern.

Utilities to create tracers:

SparseConnectivityTracer.tracer — Function

tracer(ConnectivityTracer{S}, index)
tracer(ConnectivityTracer{S}, indices)
tracer(JacobianTracer{S}, index)
tracer(JacobianTracer{S}, indices)
tracer(HessianTracer{S}, index)
tracer(HessianTracer{S}, indices)

Convenience constructor for ConnectivityTracer, JacobianTracer and HessianTracer from input indices. The provided set type S has to be an AbstractSet{<:Integer}, e.g. BitSet or Set{UInt64}.

Example

julia> tracer(JacobianTracer{BitSet}, 2)
JacobianTracer{BitSet}(2,)

julia> tracer(HessianTracer{Set{UInt64}}, 2)
HessianTracer{Set{UInt64}}(
  2 => (),
)

SparseConnectivityTracer.trace_input — Function

trace_input(T, x)
trace_input(T, x)

Enumerates input indices and constructs the specified type T of tracer. Supports ConnectivityTracer, JacobianTracer and HessianTracer.

Example

julia> x = rand(3);

julia> trace_input(ConnectivityTracer{BitSet}, x)
3-element Vector{ConnectivityTracer{BitSet}}:
 ConnectivityTracer{BitSet}(1,)
 ConnectivityTracer{BitSet}(2,)
 ConnectivityTracer{BitSet}(3,)

julia> trace_input(JacobianTracer{BitSet}, x)
3-element Vector{JacobianTracer{BitSet}}:
 JacobianTracer{BitSet}(1,)
 JacobianTracer{BitSet}(2,)
 JacobianTracer{BitSet}(3,)

julia> trace_input(HessianTracer{BitSet}, x)
3-element Vector{HessianTracer{BitSet}}:
 HessianTracer{BitSet}(
  1 => (),
)
 HessianTracer{BitSet}(
  2 => (),
)
 HessianTracer{BitSet}(
  3 => (),
)

Utility to extract input indices from tracers:

SparseConnectivityTracer.inputs — Function

inputs(tracer)

Return input indices of a ConnectivityTracer or JacobianTracer

Example

julia> a = tracer(ConnectivityTracer{BitSet}, 2)
ConnectivityTracer{BitSet}(2,)

julia> b = tracer(ConnectivityTracer{BitSet}, 4)
ConnectivityTracer{BitSet}(4,)

julia> c = a + b
ConnectivityTracer{BitSet}(2, 4)

julia> inputs(c)
2-element Vector{Int64}:
 2
 4