API Reference

SparseConnectivityTracer uses ADTypes.jl's interface for sparsity detection. In fact, the functions jacobian_sparsity and hessian_sparsity are re-exported from ADTypes.

ADTypes.jacobian_sparsity — Function

jacobian_sparsity(f, x, sd::AbstractSparsityDetector)::AbstractMatrix{Bool}
jacobian_sparsity(f!, y, x, sd::AbstractSparsityDetector)::AbstractMatrix{Bool}

Use detector sd to construct a (typically sparse) matrix S describing the pattern of nonzeroes in the Jacobian of f (resp. f!) applied at x (resp. (y, x)).

ADTypes.hessian_sparsity — Function

hessian_sparsity(f, x, sd::AbstractSparsityDetector)::AbstractMatrix{Bool}

Use detector sd to construct a (typically sparse) matrix S describing the pattern of nonzeroes in the Hessian of f applied at x.

To compute global sparsity patterns of f(x) over the entire input domain x, use

SparseConnectivityTracer.TracerSparsityDetector — Type

TracerSparsityDetector()

Global sparsity detection over the entire input domain using SparseConnectivityTracer.jl. For use with ADTypes.jl's AbstractSparsityDetector interface.

For local sparsity patterns, use TracerLocalSparsityDetector.

Keyword arguments

gradient_pattern_type::Type: Data structure used for bookkeeping of gradient sparsity patterns, used in jacobian_sparsity. Supports concrete subtypes of AbstactSet{<:Integer}. Defaults to BitSet.

hessian_pattern_type::Type: Data structure used for bookkeeping of Hessian sparsity patterns, used in hessian_sparsity. Supports concrete subtypes of AbstractDict{I, AbstractSet{I}} or AbstractSet{Tuple{I, I}}, where I <: Integer. Defaults to Dict{Int64, BitSet}.

shared_hessian_pattern::Bool: Indicate whether second-order information in Hessian sparsity patterns always shares memory and whether operators are allowed to mutate HessianTracers. Defaults to false.

If support for further pattern representations is needed, please open a feature request: https://github.com/adrhill/SparseConnectivityTracer.jl/issues

Example

julia> using SparseConnectivityTracer

julia> detector = TracerSparsityDetector()
TracerSparsityDetector()

julia> jacobian_sparsity(diff, rand(4), detector)
3×4 SparseArrays.SparseMatrixCSC{Bool, Int64} with 6 stored entries:
 1  1  ⋅  ⋅
 ⋅  1  1  ⋅
 ⋅  ⋅  1  1

julia> f(x) = x[1] + x[2]*x[3] + 1/x[4];

julia> hessian_sparsity(f, rand(4), detector)
4×4 SparseArrays.SparseMatrixCSC{Bool, Int64} with 3 stored entries:
 ⋅  ⋅  ⋅  ⋅
 ⋅  ⋅  1  ⋅
 ⋅  1  ⋅  ⋅
 ⋅  ⋅  ⋅  1

References

A. Hill and G. Dalle (2025). "Sparser, Better, Faster, Stronger: Sparsity Detection for Efficient Automatic Differentiation"

A. Hill, G. Dalle, A. Montoison (2025). "An Illustrated Guide to Automatic Sparse Differentiation"

To compute local sparsity patterns of f(x) at a specific input x, use

SparseConnectivityTracer.TracerLocalSparsityDetector — Type

TracerLocalSparsityDetector()

Local sparsity detection using SparseConnectivityTracer.jl. For use with ADTypes.jl's AbstractSparsityDetector interface.

For global sparsity patterns, use TracerSparsityDetector.

Keyword arguments

gradient_pattern_type::Type: Data structure used for bookkeeping of gradient sparsity patterns, used in jacobian_sparsity. Supports concrete subtypes of AbstactSet{<:Integer}. Defaults to BitSet.

hessian_pattern_type::Type: Data structure used for bookkeeping of Hessian sparsity patterns, used in hessian_sparsity. Supports concrete subtypes of AbstractDict{I, AbstractSet{I}} or AbstractSet{Tuple{I, I}}, where I <: Integer. Defaults to Dict{Int64, BitSet}.

shared_hessian_pattern::Bool: Indicate whether second-order information in Hessian sparsity patterns always shares memory and whether operators are allowed to mutate HessianTracers. Defaults to false.

If support for further pattern representations is needed, please open a feature request: https://github.com/adrhill/SparseConnectivityTracer.jl/issues

Example

Local sparsity patterns are less convervative than global patterns and need to be recomputed for each input x:

julia> using SparseConnectivityTracer

julia> detector = TracerLocalSparsityDetector()
TracerLocalSparsityDetector()

julia> f(x) = x[1] * x[2]; # J_f = [x[2], x[1]]

julia> jacobian_sparsity(f, [1, 0], detector)
1×2 SparseArrays.SparseMatrixCSC{Bool, Int64} with 1 stored entry:
 ⋅  1

julia> jacobian_sparsity(f, [0, 1], detector)
1×2 SparseArrays.SparseMatrixCSC{Bool, Int64} with 1 stored entry:
 1  ⋅

julia> jacobian_sparsity(f, [0, 0], detector)
1×2 SparseArrays.SparseMatrixCSC{Bool, Int64} with 0 stored entries:
 ⋅  ⋅

julia> jacobian_sparsity(f, [1, 1], detector)
1×2 SparseArrays.SparseMatrixCSC{Bool, Int64} with 2 stored entries:
 1  1

References

A. Hill and G. Dalle (2025). "Sparser, Better, Faster, Stronger: Sparsity Detection for Efficient Automatic Differentiation"

A. Hill, G. Dalle, A. Montoison (2025). "An Illustrated Guide to Automatic Sparse Differentiation"

Memory allocation

For developers requiring the allocation of output buffers that support our tracers, we additionally provide

SparseConnectivityTracer.jacobian_eltype — Function

jacobian_eltype(x, detector)

Act like eltype(x) but return the matching number type used inside Jacobian sparsity detection.

SparseConnectivityTracer.hessian_eltype — Function

hessian_eltype(x, detector)

Act like eltype(x) but return the matching number type used inside Hessian sparsity detection.

SparseConnectivityTracer.jacobian_buffer — Function

jacobian_buffer(x, detector)

Allocate a buffer similiar to x with the required tracer type for Jacobian sparsity detection. Thin wrapper around similar that doesn't expose internal types.

SparseConnectivityTracer.hessian_buffer — Function

hessian_buffer(x, detector)

Allocate a buffer similiar to x with the required tracer type for Hessian sparsity detection. Thin wrapper around similar that doesn't expose internal types.

Please note that the DiffCache from PreallocationTools.jl can be used to make caches compatible with both ForwardDiff.jl and SparseConnectivityTracer.jl.