SparseConnectivityTracer

Documentation for SparseConnectivityTracer.

API reference

SparseConnectivityTracer.Tracer

SparseConnectivityTracer.connectivity

SparseConnectivityTracer.inputs

SparseConnectivityTracer.sortedinputs

SparseConnectivityTracer.trace_input

SparseConnectivityTracer.tracer

Interface

SparseConnectivityTracer.connectivity — Function

connectivity(f, x)

Enumerates inputs x and primal outputs y=f(x) and returns sparse connectivity matrix C of size (m, n) where C[i, j] is true if the compute graph connects the i-th entry in y to the j-th entry in x.

Example

julia> x = rand(3);

julia> f(x) = [x[1]^2, 2 * x[1] * x[2]^2, sin(x[3])];

julia> connectivity(f, x)
3×3 SparseArrays.SparseMatrixCSC{Bool, UInt64} with 4 stored entries:
 1  ⋅  ⋅
 1  1  ⋅
 ⋅  ⋅  1

connectivity(f!, y, x)

Enumerates inputs x and primal outputs y after f!(y, x) and returns sparse connectivity matrix C of size (m, n) where C[i, j] is true if the compute graph connects the i-th entry in y to the j-th entry in x.

Internals

SparseConnectivityTracer works by pushing a Number type called Tracer through generic functions:

SparseConnectivityTracer.Tracer — Type

Tracer(indexset) <: Number

Number type keeping track of input indices of previous computations.

See also the convenience constructor tracer. For a higher-level interface, refer to connectivity.

Examples

By enumerating inputs with tracers, we can keep track of input connectivities:

julia> xt = [tracer(1), tracer(2), tracer(3)]
3-element Vector{Tracer}:
 Tracer(1,)
 Tracer(2,)
 Tracer(3,)

julia> f(x) = [x[1]^2, 2 * x[1] * x[2]^2, sin(x[3])];

julia> yt = f(xt)
3-element Vector{Tracer}:
   Tracer(1,)
 Tracer(1, 2)
   Tracer(3,)

This works by overloading operators to either keep input connectivities constant, compute unions or set connectivities to zero:

julia> x = tracer(1, 2, 3)
Tracer(1, 2, 3)

julia> sin(x)  # Most operators don't modify input connectivities.
Tracer(1, 2, 3)

julia> 2 * x^3
Tracer(1, 2, 3)

julia> zero(x) # Tracer is strictly operator overloading... 
Tracer()

julia> 0 * x   # ...and doesn't look at input values.
Tracer(1, 2, 3)

julia> y = tracer(3, 5)
Tracer(3, 5)

julia> x + y   # Operations on two Tracers construct union sets
Tracer(1, 2, 3, 5)

julia> x ^ y
Tracer(1, 2, 3, 5)

Tracer also supports random number generation and pre-allocations:

julia> M = rand(Tracer, 3, 2)
3×2 Matrix{Tracer}:
 Tracer()  Tracer()
 Tracer()  Tracer()
 Tracer()  Tracer()

julia> similar(M)
3×2 Matrix{Tracer}:
 Tracer()  Tracer()
 Tracer()  Tracer()
 Tracer()  Tracer()

julia> M * [x, y]
3-element Vector{Tracer}:
 Tracer(1, 2, 3, 5)
 Tracer(1, 2, 3, 5)
 Tracer(1, 2, 3, 5)

SparseConnectivityTracer.tracer — Function

tracer(index)
tracer(indices)

Convenience constructor for Tracer from input indices.

SparseConnectivityTracer.trace_input — Function

trace_input(x)

Enumerates input indices and constructs Tracers.

Example

julia> x = rand(3);

julia> f(x) = [x[1]^2, 2 * x[1] * x[2]^2, sin(x[3])];

julia> xt = trace_input(x)
3-element Vector{Tracer}:
 Tracer(1,)
 Tracer(2,)
 Tracer(3,)

julia> yt = f(xt)
3-element Vector{Tracer}:
   Tracer(1,)
 Tracer(1, 2)
   Tracer(3,)

The following utilities can be used to extract input indices from Tracers:

SparseConnectivityTracer.inputs — Function

inputs(tracer)

Return raw UInt64 input indices of a Tracer. See also sortedinputs.

Example

julia> t = tracer(1, 2, 4)
Tracer(1, 2, 4)

julia> inputs(t)
3-element Vector{UInt64}:
 0x0000000000000004
 0x0000000000000002
 0x0000000000000001

SparseConnectivityTracer.sortedinputs — Function

sortedinputs(tracer)
sortedinputs([T=Int], tracer)

Return sorted input indices of a Tracer. See also inputs.

Example

julia> t = tracer(1, 2, 4)
Tracer(1, 2, 4)

julia> sortedinputs(t)
3-element Vector{Int64}:
 1
 2
 4

julia> sortedinputs(UInt8, t)
3-element Vector{UInt8}:
 0x01
 0x02
 0x04