API Reference

SparseConnectivityTracer.ConnectivityTracer

SparseConnectivityTracer.HessianTracer

SparseConnectivityTracer.JacobianTracer

SparseConnectivityTracer.TracerSparsityDetector

SparseConnectivityTracer.inputs

SparseConnectivityTracer.pattern

SparseConnectivityTracer.trace_input

SparseConnectivityTracer.tracer

Interface

SparseConnectivityTracer.pattern — Function

pattern(f, ConnectivityTracer, x)

Enumerates inputs x and primal outputs y = f(x) and returns sparse matrix C of size (m, n) where C[i, j] is true if the compute graph connects the i-th entry in y to the j-th entry in x.

pattern(f, JacobianTracer, x)

Computes the sparsity pattern of the Jacobian of y = f(x).

pattern(f, HessianTracer, x)

Computes the sparsity pattern of the Hessian of a scalar function y = f(x).

Examples

First order

julia> x = rand(3);

julia> f(x) = [x[1]^2, 2 * x[1] * x[2]^2, sin(x[3])];

julia> pattern(f, ConnectivityTracer, x)
3×3 SparseArrays.SparseMatrixCSC{Bool, UInt64} with 4 stored entries:
 1  ⋅  ⋅
 1  1  ⋅
 ⋅  ⋅  1

Second order

julia> x = rand(5);

julia> f(x) = x[1] + x[2]*x[3] + 1/x[4] + 1*x[5];

julia> pattern(f, HessianTracer, x)
5×5 SparseArrays.SparseMatrixCSC{Bool, UInt64} with 3 stored entries:
 ⋅  ⋅  ⋅  ⋅  ⋅
 ⋅  ⋅  1  ⋅  ⋅
 ⋅  1  ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅  1  ⋅
 ⋅  ⋅  ⋅  ⋅  ⋅

julia> g(x) = f(x) + x[2]^x[5];

julia> pattern(g, HessianTracer, x)
5×5 SparseArrays.SparseMatrixCSC{Bool, UInt64} with 7 stored entries:
 ⋅  ⋅  ⋅  ⋅  ⋅
 ⋅  1  1  ⋅  1
 ⋅  1  ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅  1  ⋅
 ⋅  1  ⋅  ⋅  1

pattern(f!, y, JacobianTracer, x)

Computes the sparsity pattern of the Jacobian of f!(y, x).

pattern(f!, y, ConnectivityTracer, x)

Enumerates inputs x and primal outputs y after f!(y, x) and returns sparse matrix C of size (m, n) where C[i, j] is true if the compute graph connects the i-th entry in y to the j-th entry in x.

SparseConnectivityTracer.TracerSparsityDetector — Type

TracerSparsityDetector <: ADTypes.AbstractSparsityDetector

Singleton struct for integration with the sparsity detection framework of ADTypes.jl.

Example

julia> using ADTypes, SparseConnectivityTracer

julia> ADTypes.jacobian_sparsity(diff, rand(4), TracerSparsityDetector())
3×4 SparseArrays.SparseMatrixCSC{Bool, UInt64} with 6 stored entries:
 1  1  ⋅  ⋅
 ⋅  1  1  ⋅
 ⋅  ⋅  1  1

julia> using ADTypes, SparseConnectivityTracer

julia> f(x) = x[1] + x[2]*x[3] + 1/x[4];

julia> ADTypes.hessian_sparsity(f, rand(4), TracerSparsityDetector())
4×4 SparseArrays.SparseMatrixCSC{Bool, UInt64} with 3 stored entries:
 ⋅  ⋅  ⋅  ⋅
 ⋅  ⋅  1  ⋅
 ⋅  1  ⋅  ⋅
 ⋅  ⋅  ⋅  1

Internals

SparseConnectivityTracer works by pushing Number types called tracers through generic functions. Currently, two tracer types are provided:

SparseConnectivityTracer.ConnectivityTracer — Type

ConnectivityTracer(indexset) <: Number

Number type keeping track of input indices of previous computations.

See also the convenience constructor tracer. For a higher-level interface, refer to pattern.

SparseConnectivityTracer.JacobianTracer — Type

JacobianTracer(indexset) <: Number

Number type keeping track of input indices of previous computations with non-zero derivatives.

See also the convenience constructor tracer. For a higher-level interface, refer to pattern.

SparseConnectivityTracer.HessianTracer — Type

HessianTracer(indexset) <: Number

Number type keeping track of input indices of previous computations with non-zero first and second derivatives.

See also the convenience constructor tracer. For a higher-level interface, refer to pattern.

Utilities to create tracers:

SparseConnectivityTracer.tracer — Function

tracer(JacobianTracer, index)
tracer(JacobianTracer, indices)
tracer(ConnectivityTracer, index)
tracer(ConnectivityTracer, indices)

Convenience constructor for JacobianTracer ConnectivityTracer from input indices.

SparseConnectivityTracer.trace_input — Function

trace_input(JacobianTracer, x)
trace_input(ConnectivityTracer, x)

Enumerates input indices and constructs the specified type of tracer. Supports JacobianTracer and ConnectivityTracer.

Example

julia> x = rand(3);

julia> trace_input(ConnectivityTracer, x)
3-element Vector{ConnectivityTracer}:
 ConnectivityTracer(1,)
 ConnectivityTracer(2,)
 ConnectivityTracer(3,)

julia> trace_input(JacobianTracer, x)
3-element Vector{JacobianTracer}:
 JacobianTracer(1,)
 JacobianTracer(2,)
 JacobianTracer(3,)

julia> trace_input(HessianTracer, x)
3-element Vector{HessianTracer}:
 HessianTracer(
  1 => (),
)
 HessianTracer(
  2 => (),
)
 HessianTracer(
  3 => (),
)

Utility to extract input indices from tracers:

SparseConnectivityTracer.inputs — Function

inputs(tracer)

Return raw UInt64 input indices of a ConnectivityTracer or JacobianTracer

Example

julia> t = tracer(ConnectivityTracer, 1, 2, 4)
ConnectivityTracer(1, 2, 4)

julia> inputs(t)
3-element Vector{Int64}:
 1
 2
 4