API Reference

SparseConnectivityTracer.ConnectivityTracer

SparseConnectivityTracer.Dual

SparseConnectivityTracer.DuplicateVector

SparseConnectivityTracer.GradientTracer

SparseConnectivityTracer.HessianTracer

SparseConnectivityTracer.RecursiveSet

SparseConnectivityTracer.SortedVector

SparseConnectivityTracer.TracerLocalSparsityDetector

SparseConnectivityTracer.TracerSparsityDetector

SparseConnectivityTracer.connectivity_pattern

SparseConnectivityTracer.hessian_pattern

SparseConnectivityTracer.jacobian_pattern

SparseConnectivityTracer.local_connectivity_pattern

SparseConnectivityTracer.local_hessian_pattern

SparseConnectivityTracer.local_jacobian_pattern

ADTypes Interface

For package developers, we recommend using the ADTypes.jl interface.

To compute global sparsity patterns of f(x) over the entire input domain x, use

SparseConnectivityTracer.TracerSparsityDetector — Type

TracerSparsityDetector <: ADTypes.AbstractSparsityDetector

Singleton struct for integration with the sparsity detection framework of ADTypes.jl.

Computes global sparsity patterns over the entire input domain. For local sparsity patterns at a specific input point, use TracerLocalSparsityDetector.

Example

julia> using ADTypes, SparseConnectivityTracer

julia> ADTypes.jacobian_sparsity(diff, rand(4), TracerSparsityDetector())
3×4 SparseArrays.SparseMatrixCSC{Bool, Int64} with 6 stored entries:
 1  1  ⋅  ⋅
 ⋅  1  1  ⋅
 ⋅  ⋅  1  1

julia> using ADTypes, SparseConnectivityTracer

julia> f(x) = x[1] + x[2]*x[3] + 1/x[4];

julia> ADTypes.hessian_sparsity(f, rand(4), TracerSparsityDetector())
4×4 SparseArrays.SparseMatrixCSC{Bool, Int64} with 3 stored entries:
 ⋅  ⋅  ⋅  ⋅
 ⋅  ⋅  1  ⋅
 ⋅  1  ⋅  ⋅
 ⋅  ⋅  ⋅  1

To compute local sparsity patterns of f(x) at a specific input x, use

SparseConnectivityTracer.TracerLocalSparsityDetector — Type

TracerLocalSparsityDetector <: ADTypes.AbstractSparsityDetector

Singleton struct for integration with the sparsity detection framework of ADTypes.jl.

Computes local sparsity patterns at an input point x. For global sparsity patterns, use TracerSparsityDetector.

Example

julia> using ADTypes, SparseConnectivityTracer

julia> f(x) = x[1] > x[2] ? x[1:3] : x[2:4];

julia> ADTypes.jacobian_sparsity(f, [1.0, 2.0, 3.0, 4.0], TracerLocalSparsityDetector())
3×4 SparseArrays.SparseMatrixCSC{Bool, Int64} with 3 stored entries:
 ⋅  1  ⋅  ⋅
 ⋅  ⋅  1  ⋅
 ⋅  ⋅  ⋅  1

julia> ADTypes.jacobian_sparsity(f, [2.0, 1.0, 3.0, 4.0], TracerLocalSparsityDetector())
3×4 SparseArrays.SparseMatrixCSC{Bool, Int64} with 3 stored entries:
 1  ⋅  ⋅  ⋅
 ⋅  1  ⋅  ⋅
 ⋅  ⋅  1  ⋅

julia> using ADTypes, SparseConnectivityTracer

julia> f(x) = x[1] + max(x[2], x[3]) * x[3] + 1/x[4];

julia> ADTypes.hessian_sparsity(f, [1.0, 2.0, 3.0, 4.0], TracerLocalSparsityDetector())
4×4 SparseArrays.SparseMatrixCSC{Bool, Int64} with 2 stored entries:
 ⋅  ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅  ⋅
 ⋅  ⋅  1  ⋅
 ⋅  ⋅  ⋅  1

Legacy Interface

Global sparsity

The following functions can be used to compute global sparsity patterns of f(x) over the entire input domain x.

SparseConnectivityTracer.connectivity_pattern — Function

connectivity_pattern(f, x)
connectivity_pattern(f, x, T)

Enumerates inputs x and primal outputs y = f(x) and returns sparse matrix C of size (m, n) where C[i, j] is true if the compute graph connects the i-th entry in y to the j-th entry in x.

The type of ConnectivityTracer can be specified as an optional argument and defaults to SparseConnectivityTracer.ConnectivityTracer{BitSet}.

Example

julia> x = rand(3);

julia> f(x) = [x[1]^2, 2 * x[1] * x[2]^2, sign(x[3])];

julia> connectivity_pattern(f, x)
3×3 SparseArrays.SparseMatrixCSC{Bool, Int64} with 4 stored entries:
 1  ⋅  ⋅
 1  1  ⋅
 ⋅  ⋅  1

connectivity_pattern(f!, y, x)
connectivity_pattern(f!, y, x, T)

Enumerates inputs x and primal outputs y after f!(y, x) and returns sparse matrix C of size (m, n) where C[i, j] is true if the compute graph connects the i-th entry in y to the j-th entry in x.

The type of ConnectivityTracer can be specified as an optional argument and defaults to SparseConnectivityTracer.ConnectivityTracer{BitSet}.

SparseConnectivityTracer.jacobian_pattern — Function

jacobian_pattern(f, x)
jacobian_pattern(f, x, T)

Compute the sparsity pattern of the Jacobian of y = f(x).

The type of GradientTracer can be specified as an optional argument and defaults to SparseConnectivityTracer.ConnectivityTracer{BitSet}.

Example

julia> x = rand(3);

julia> f(x) = [x[1]^2, 2 * x[1] * x[2]^2, sign(x[3])];

julia> jacobian_pattern(f, x)
3×3 SparseArrays.SparseMatrixCSC{Bool, Int64} with 3 stored entries:
 1  ⋅  ⋅
 1  1  ⋅
 ⋅  ⋅  ⋅

jacobian_pattern(f!, y, x)
jacobian_pattern(f!, y, x, T)

Compute the sparsity pattern of the Jacobian of f!(y, x).

The type of GradientTracer can be specified as an optional argument and defaults to SparseConnectivityTracer.GradientTracer{BitSet}.

SparseConnectivityTracer.hessian_pattern — Function

hessian_pattern(f, x)
hessian_pattern(f, x, T)

Computes the sparsity pattern of the Hessian of a scalar function y = f(x).

The type of HessianTracer can be specified as an optional argument and defaults to SparseConnectivityTracer.HessianTracer{BitSet, Set{Tuple{Int64, Int64}}}.

Example

julia> x = rand(5);

julia> f(x) = x[1] + x[2]*x[3] + 1/x[4] + 1*x[5];

julia> hessian_pattern(f, x)
5×5 SparseArrays.SparseMatrixCSC{Bool, Int64} with 3 stored entries:
 ⋅  ⋅  ⋅  ⋅  ⋅
 ⋅  ⋅  1  ⋅  ⋅
 ⋅  1  ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅  1  ⋅
 ⋅  ⋅  ⋅  ⋅  ⋅

julia> g(x) = f(x) + x[2]^x[5];

julia> hessian_pattern(g, x)
5×5 SparseArrays.SparseMatrixCSC{Bool, Int64} with 7 stored entries:
 ⋅  ⋅  ⋅  ⋅  ⋅
 ⋅  1  1  ⋅  1
 ⋅  1  ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅  1  ⋅
 ⋅  1  ⋅  ⋅  1

TracerSparsityDetector is the ADTypes equivalent of these functions.

Local sparsity

The following functions can be used to compute local sparsity patterns of f(x) at a specific input x. Note that these patterns are sparser than global patterns but need to be recomputed when x changes.

SparseConnectivityTracer.local_connectivity_pattern — Function

local_connectivity_pattern(f, x)
local_connectivity_pattern(f, x, P)

Enumerates inputs x and primal outputs y = f(x) and returns sparse matrix C of size (m, n) where C[i, j] is true if the compute graph connects the i-th entry in y to the j-th entry in x.

Unlike connectivity_pattern, this function supports control flow and comparisons.

The type of ConnectivityTracer can be specified as an optional argument and defaults to SparseConnectivityTracer.ConnectivityTracer{BitSet}.

Example

julia> f(x) = ifelse(x[2] < x[3], x[1] + x[2], x[3] * x[4]);

julia> x = [1 2 3 4];

julia> local_connectivity_pattern(f, x)
1×4 SparseArrays.SparseMatrixCSC{Bool, Int64} with 2 stored entries:
 1  1  ⋅  ⋅

julia> x = [1 3 2 4];

julia> local_connectivity_pattern(f, x)
1×4 SparseArrays.SparseMatrixCSC{Bool, Int64} with 2 stored entries:
 ⋅  ⋅  1  1

local_connectivity_pattern(f!, y, x)
local_connectivity_pattern(f!, y, x, T)

Enumerates inputs x and primal outputs y after f!(y, x) and returns sparse matrix C of size (m, n) where C[i, j] is true if the compute graph connects the i-th entry in y to the j-th entry in x.

Unlike connectivity_pattern, this function supports control flow and comparisons.

The type of ConnectivityTracer can be specified as an optional argument and defaults to SparseConnectivityTracer.ConnectivityTracer{BitSet}.

SparseConnectivityTracer.local_jacobian_pattern — Function

local_jacobian_pattern(f, x)
local_jacobian_pattern(f, x, T)

Compute the local sparsity pattern of the Jacobian of y = f(x) at x.

The type of GradientTracer can be specified as an optional argument and defaults to SparseConnectivityTracer.GradientTracer{BitSet}.

Example

julia> x = [1.0, 2.0, 3.0];

julia> f(x) = [x[1]^2, 2 * x[1] * x[2]^2, max(x[2],x[3])];

julia> local_jacobian_pattern(f, x)
3×3 SparseArrays.SparseMatrixCSC{Bool, Int64} with 4 stored entries:
 1  ⋅  ⋅
 1  1  ⋅
 ⋅  ⋅  1

local_jacobian_pattern(f!, y, x)
local_jacobian_pattern(f!, y, x, T)

Compute the local sparsity pattern of the Jacobian of f!(y, x) at x.

The type of GradientTracer can be specified as an optional argument and defaults to SparseConnectivityTracer.GradientTracer{BitSet}.

SparseConnectivityTracer.local_hessian_pattern — Function

local_hessian_pattern(f, x)
local_hessian_pattern(f, x, T)

Computes the local sparsity pattern of the Hessian of a scalar function y = f(x) at x.

The type of HessianTracer can be specified as an optional argument and defaults to SparseConnectivityTracer.HessianTracer{BitSet, Set{Tuple{Int64, Int64}}}.

Example

julia> x = [1.0 3.0 5.0 1.0 2.0];

julia> f(x) = x[1] + x[2]*x[3] + 1/x[4] + x[2] * max(x[1], x[5]);

julia> local_hessian_pattern(f, x)
5×5 SparseArrays.SparseMatrixCSC{Bool, Int64} with 5 stored entries:
 ⋅  ⋅  ⋅  ⋅  ⋅
 ⋅  ⋅  1  ⋅  1
 ⋅  1  ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅  1  ⋅
 ⋅  1  ⋅  ⋅  ⋅

julia> x = [4.0 3.0 5.0 1.0 2.0];

julia> local_hessian_pattern(f, x)
5×5 SparseArrays.SparseMatrixCSC{Bool, Int64} with 5 stored entries:
 ⋅  1  ⋅  ⋅  ⋅
 1  ⋅  1  ⋅  ⋅
 ⋅  1  ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅  1  ⋅
 ⋅  ⋅  ⋅  ⋅  ⋅

TracerLocalSparsityDetector is the ADTypes equivalent of these functions.

Internals

Warning

Internals may change without warning in a future release of SparseConnectivityTracer.

SparseConnectivityTracer works by pushing Real number types called tracers through generic functions. Currently, three tracer types are provided:

SparseConnectivityTracer.ConnectivityTracer — Type

struct ConnectivityTracer{I} <: SparseConnectivityTracer.AbstractTracer

Real number type keeping track of input indices of previous computations.

For a higher-level interface, refer to connectivity_pattern.

Fields

inputs::Any: Sparse representation of connected inputs.

isempty::Bool: Indicator whether pattern in tracer contains only zeros.

SparseConnectivityTracer.GradientTracer — Type

struct GradientTracer{G} <: SparseConnectivityTracer.AbstractTracer

Real number type keeping track of non-zero gradient entries.

For a higher-level interface, refer to jacobian_pattern.

Fields

gradient::Any: Sparse representation of non-zero entries in the gradient.

isempty::Bool: Indicator whether gradient in tracer contains only zeros.

SparseConnectivityTracer.HessianTracer — Type

struct HessianTracer{G, H} <: SparseConnectivityTracer.AbstractTracer

Real number type keeping track of non-zero gradient and Hessian entries.

For a higher-level interface, refer to hessian_pattern.

Fields

gradient::Any: Sparse representation of non-zero entries in the gradient and the Hessian.

hessian::Any: Sparse representation of non-zero entries in the Hessian.

isempty::Bool: Indicator whether gradient and Hessian in tracer both contain only zeros.

These can be used alone or inside of the dual number type Dual, which keeps track of the primal computation and allows tracing through comparisons and control flow:

SparseConnectivityTracer.Dual — Type

struct Dual{P<:Real, T<:SparseConnectivityTracer.AbstractTracer} <: Real

Dual Real number type keeping track of the results of a primal computation as well as a tracer.

Fields

primal::Real

tracer::SparseConnectivityTracer.AbstractTracer

We also define alternative pseudo-set types that can deliver faster union:

SparseConnectivityTracer.DuplicateVector — Type

DuplicateVector

Vector that can have duplicate values, for which union is just concatenation.

SparseConnectivityTracer.RecursiveSet — Type

RecursiveSet

Lazy union of sets.

SparseConnectivityTracer.SortedVector — Type

SortedVector

Sorted vector without duplicates, designed for fast set unions with merging.