API Reference

SparseConnectivityTracer.ConnectivityTracer

SparseConnectivityTracer.DuplicateVector

SparseConnectivityTracer.HessianTracer

SparseConnectivityTracer.JacobianTracer

SparseConnectivityTracer.RecursiveSet

SparseConnectivityTracer.SortedVector

SparseConnectivityTracer.TracerSparsityDetector

SparseConnectivityTracer.connectivity_pattern

SparseConnectivityTracer.hessian_pattern

SparseConnectivityTracer.jacobian_pattern

Interface

SparseConnectivityTracer.connectivity_pattern — Function

connectivity_pattern(f, x)
connectivity_pattern(f, x, T)

Enumerates inputs x and primal outputs y = f(x) and returns sparse matrix C of size (m, n) where C[i, j] is true if the compute graph connects the i-th entry in y to the j-th entry in x.

The type of index set S can be specified as an optional argument and defaults to BitSet.

Example

julia> x = rand(3);

julia> f(x) = [x[1]^2, 2 * x[1] * x[2]^2, sign(x[3])];

julia> connectivity_pattern(f, x)
3×3 SparseArrays.SparseMatrixCSC{Bool, UInt64} with 4 stored entries:
 1  ⋅  ⋅
 1  1  ⋅
 ⋅  ⋅  1

connectivity_pattern(f!, y, x)
connectivity_pattern(f!, y, x, T)

Enumerates inputs x and primal outputs y after f!(y, x) and returns sparse matrix C of size (m, n) where C[i, j] is true if the compute graph connects the i-th entry in y to the j-th entry in x.

The type of index set S can be specified as an optional argument and defaults to BitSet.

SparseConnectivityTracer.jacobian_pattern — Function

jacobian_pattern(f, x)
jacobian_pattern(f, x, T)

Compute the sparsity pattern of the Jacobian of y = f(x).

The type of index set S can be specified as an optional argument and defaults to BitSet.

Example

julia> x = rand(3);

julia> f(x) = [x[1]^2, 2 * x[1] * x[2]^2, sign(x[3])];

julia> jacobian_pattern(f, x)
3×3 SparseArrays.SparseMatrixCSC{Bool, UInt64} with 3 stored entries:
 1  ⋅  ⋅
 1  1  ⋅
 ⋅  ⋅  ⋅

jacobian_pattern(f!, y, x)
jacobian_pattern(f!, y, x, T)

Compute the sparsity pattern of the Jacobian of f!(y, x).

The type of index set S can be specified as an optional argument and defaults to BitSet.

SparseConnectivityTracer.hessian_pattern — Function

hessian_pattern(f, x)
hessian_pattern(f, x, T)

Computes the sparsity pattern of the Hessian of a scalar function y = f(x).

The type of index set S can be specified as an optional argument and defaults to BitSet.

Example

julia> x = rand(5);

julia> f(x) = x[1] + x[2]*x[3] + 1/x[4] + 1*x[5];

julia> hessian_pattern(f, x)
5×5 SparseArrays.SparseMatrixCSC{Bool, UInt64} with 3 stored entries:
 ⋅  ⋅  ⋅  ⋅  ⋅
 ⋅  ⋅  1  ⋅  ⋅
 ⋅  1  ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅  1  ⋅
 ⋅  ⋅  ⋅  ⋅  ⋅

julia> g(x) = f(x) + x[2]^x[5];

julia> hessian_pattern(g, x)
5×5 SparseArrays.SparseMatrixCSC{Bool, UInt64} with 7 stored entries:
 ⋅  ⋅  ⋅  ⋅  ⋅
 ⋅  1  1  ⋅  1
 ⋅  1  ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅  1  ⋅
 ⋅  1  ⋅  ⋅  1

SparseConnectivityTracer.TracerSparsityDetector — Type

TracerSparsityDetector <: ADTypes.AbstractSparsityDetector

Singleton struct for integration with the sparsity detection framework of ADTypes.jl.

Example

julia> using ADTypes, SparseConnectivityTracer

julia> ADTypes.jacobian_sparsity(diff, rand(4), TracerSparsityDetector())
3×4 SparseArrays.SparseMatrixCSC{Bool, UInt64} with 6 stored entries:
 1  1  ⋅  ⋅
 ⋅  1  1  ⋅
 ⋅  ⋅  1  1

julia> using ADTypes, SparseConnectivityTracer

julia> f(x) = x[1] + x[2]*x[3] + 1/x[4];

julia> ADTypes.hessian_sparsity(f, rand(4), TracerSparsityDetector())
4×4 SparseArrays.SparseMatrixCSC{Bool, UInt64} with 3 stored entries:
 ⋅  ⋅  ⋅  ⋅
 ⋅  ⋅  1  ⋅
 ⋅  1  ⋅  ⋅
 ⋅  ⋅  ⋅  1

Internals

SparseConnectivityTracer works by pushing Number types called tracers through generic functions. Currently, three tracer types are provided:

SparseConnectivityTracer.ConnectivityTracer — Type

ConnectivityTracer{I,S}(indexset) <: Number

Number type keeping track of input indices of previous computations.

The provided index set type S has to satisfy the following conditions:

it is an iterable with <:Integer element type

it implements union

Subtypes of AbstractSet{<:Integer} are a natural choice, like BitSet or Set{UInt64}.

For a higher-level interface, refer to connectivity_pattern.

SparseConnectivityTracer.JacobianTracer — Type

JacobianTracer{I,S}(indexset) <: Number

Number type keeping track of input indices of previous computations with non-zero derivatives.

The provided index set type S has to satisfy the following conditions:

it is an iterable with <:Integer element type

it implements union

Subtypes of AbstractSet{<:Integer} are a natural choice, like BitSet or Set{UInt64}.

For a higher-level interface, refer to jacobian_pattern.

SparseConnectivityTracer.HessianTracer — Type

HessianTracer{I,S,D}(indexset) <: Number

Number type keeping track of input indices of previous computations with non-zero first and second derivatives.

The provided index set type S has to satisfy the following conditions:

it is an iterable with <:Integer element type

it implements union

Subtypes of AbstractSet{<:Integer} are a natural choice, like BitSet or Set{UInt64}.

For a higher-level interface, refer to hessian_pattern.

We also define alternative pseudo-set types that can deliver faster union:

SparseConnectivityTracer.DuplicateVector — Type

DuplicateVector

Vector that can have duplicate values, for which union is just concatenation.

SparseConnectivityTracer.RecursiveSet — Type

RecursiveSet

Lazy union of sets.

SparseConnectivityTracer.SortedVector — Type

SortedVector

Sorted vector without duplicates, designed for fast set unions with merging.